# optimize(lb): WRR reduces memory usage to minimum positive period