--- a/pkg/loadbalance/weighted_balancer_test.go
+++ b/pkg/loadbalance/weighted_balancer_test.go
@@ -275,6 +275,42 @@ func BenchmarkGetPicker(b *testing.B) {
 	}
 }
 
+func BenchmarkWeightedPicker_Next(b *testing.B) {
+	ctx := context.Background()
+
+	makeNInstances := func(n, base int, seq []int) (res []discovery.Instance) {
+		for i := 0; i < n; i++ {
+			weight := seq[i%len(seq)] * base
+			res = append(res, discovery.NewInstance("tcp", fmt.Sprintf("addr[%d]-weight[%d]", i, weight), weight, nil))
+		}
+		return
+	}
+
+	for _, tc := range balancerTestcases {
+		b.Run(tc.Name, func(b *testing.B) {
+			balancer := tc.factory()
+
+			n := 10
+			for i := 0; i < 4; i++ {
+				b.Run(fmt.Sprintf("%dins", n), func(b *testing.B) {
+					inss := makeNInstances(n, 1000, []int{1, 2, 3, 4, 5, 6, 7, 8, 9, 10})
+					e := discovery.Result{
+						Cacheable: true,
+						CacheKey:  "test",
+						Instances: inss,
+					}
+					picker := balancer.GetPicker(e)
+					b.ResetTimer()
+					for i := 0; i < b.N; i++ {
+						picker.Next(ctx, nil)
+					}
+				})
+				n *= 10
+			}
+		})
+	}
+}
+
 func genInstList() []discovery.Instance {
 	n := 1000
 	insList := make([]discovery.Instance, 0, n)