@jizhuozhi 2023-07-04
Hello, @YangruiEmma . Here is a simple benchmark which makes a very large cluster with very high weights. There is no `StopTimer()` because I want to record the memory allocation. (Because it's not reusable, so I would not commit it).
```go
func BenchmarkWeightedRoundRobinPicker_Next(b *testing.B) {
	ctx := context.Background()

	makeNInstances := func(n, base int, seq []int) (res []discovery.Instance) {
		for i := 0; i < n; i++ {
			weight := seq[i%len(seq)] * base
			res = append(res, discovery.NewInstance("tcp", fmt.Sprintf("addr[%d]-weight[%d]", i, weight), weight, nil))
		}
		return
	}

	n := 10
	for i := 0; i < 4; i++ {
		b.Run(fmt.Sprintf("%d-instances", n), func(b *testing.B) {
			inss := makeNInstances(n, 1000, []int{1, 2, 3, 4, 5, 6, 7, 8, 9, 10})
			picker := newWeightedRoundRobinPicker(inss)
			for i := 0; i < b.N; i++ {
				picker.Next(ctx, nil)
			}
		})
		n *= 10
	}
}
```

Benchmark result:

Before
```
goos: darwin
goarch: arm64
pkg: github.com/cloudwego/kitex/pkg/loadbalance
BenchmarkWeightedRoundRobinPicker_Next
BenchmarkWeightedRoundRobinPicker_Next/10-instances
BenchmarkWeightedRoundRobinPicker_Next/10-instances-12         	85309125	        13.98 ns/op	       0 B/op	       0 allocs/op
BenchmarkWeightedRoundRobinPicker_Next/100-instances
BenchmarkWeightedRoundRobinPicker_Next/100-instances-12        	68167687	        15.88 ns/op	       0 B/op	       0 allocs/op
BenchmarkWeightedRoundRobinPicker_Next/1000-instances
BenchmarkWeightedRoundRobinPicker_Next/1000-instances-12       	  453632	      2369 ns/op	     194 B/op	       0 allocs/op
BenchmarkWeightedRoundRobinPicker_Next/10000-instances
BenchmarkWeightedRoundRobinPicker_Next/10000-instances-12      	   42087	     28360 ns/op	   20957 B/op	       1 allocs/op
PASS
```

After
```
goos: darwin
goarch: arm64
pkg: github.com/cloudwego/kitex/pkg/loadbalance
BenchmarkWeightedRoundRobinPicker_Next
BenchmarkWeightedRoundRobinPicker_Next/10-instances
BenchmarkWeightedRoundRobinPicker_Next/10-instances-12         	89161636	        13.19 ns/op	       0 B/op	       0 allocs/op
BenchmarkWeightedRoundRobinPicker_Next/100-instances
BenchmarkWeightedRoundRobinPicker_Next/100-instances-12        	89877817	        13.18 ns/op	       0 B/op	       0 allocs/op
BenchmarkWeightedRoundRobinPicker_Next/1000-instances
BenchmarkWeightedRoundRobinPicker_Next/1000-instances-12       	89288391	        13.34 ns/op	       0 B/op	       0 allocs/op
BenchmarkWeightedRoundRobinPicker_Next/10000-instances
BenchmarkWeightedRoundRobinPicker_Next/10000-instances-12      	   41743	     27685 ns/op	      70 B/op	       1 allocs/op
PASS
```

The "ns/op" is averaged, `buildVirtualWrrNodes` completed earlier, the value is smaller. The total memory usage is "B/op" * "ops".