@jizhuozhi 2023-07-06
> Could you please add your benchmark in test cases?

Done